Memfault: add new POST API and reduce memory usage
Add new POST V2 API. This API only opens the socket
if there is data to send and gives the user control of the
chunk size.

Change file save API to give the user control of chunk size. This lets a user of POST and file save use the same
data buffer.

bug 21921